Address YAMLLoadWarning: calling yaml.load() without Loader=... is deprecated
PyYaml is switching to a safer mechanism for loading yaml (which avoids python injection attacks). The fix is straight-forward and described [here](https://github.com/yaml/pyyaml/wiki/PyYAML-yaml.load(input)-Deprecation).

The `FullLoader` is called automatically after the warning is issued so there will be NO change in functionality.