Preserve trailing newline at end of file
jinja2 does **not** preserve trailing newlines by default which is a weird choice for a template engine. Given that the purpose of j2cli_3 is to work as a command-line template renderer it makes sense for it to **not** make changes that are not explicitly specified.

The current behavior introduced a difficult to diagnose bug in my usage.